backend/app_backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2==2.9.5
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
